# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GovernedWorkflowError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "anyOf": [
                  {
                    "items": {},
                    "type": "array"
                  },
                  {
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "details",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object",
        "x-workflow-error-statuses": {
          "workflow_authentication_required": 401,
          "workflow_conflict": 409,
          "workflow_disabled": 503,
          "workflow_forbidden": 403,
          "workflow_invalid_input": 422,
          "workflow_invalid_result": 502,
          "workflow_mutation_unknown": 503,
          "workflow_receipt_expired": 409,
          "workflow_receipt_future": 409,
          "workflow_receipt_unverified": 403,
          "workflow_request_too_large": 413,
          "workflow_service_failed": 503,
          "workflow_terminal": 409,
          "workflow_unavailable": 503
        }
      },
      "GovernedWorkflowMutation": {
        "additionalProperties": false,
        "properties": {
          "action_digest": {
            "pattern": "^(?:sha256:[0-9a-f]{64})$(?![\\s\\S])",
            "type": "string"
          },
          "expected_state_digest": {
            "pattern": "^(?:sha256:[0-9a-f]{64})$(?![\\s\\S])",
            "type": "string"
          },
          "request_id": {
            "pattern": "^(?:req_[0-9a-f]{32})$(?![\\s\\S])",
            "type": "string"
          },
          "run_id": {
            "pattern": "^(?:run_[0-9a-f]{32})$(?![\\s\\S])",
            "type": "string"
          },
          "schema_version": {
            "const": "openmed.service.workflow_request.v1",
            "type": "string"
          },
          "workflow_id": {
            "maxLength": 512,
            "pattern": "^(?:workflow:(?=[^/]{1,253}/)(?:[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+)/(?:[a-z][a-z0-9-]{0,63})(?:@(?:(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)\\.(?:0|[1-9][0-9]*)))?)$(?![\\s\\S])",
            "type": "string"
          }
        },
        "required": [
          "action_digest",
          "expected_state_digest",
          "request_id",
          "run_id",
          "schema_version",
          "workflow_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "WorkflowApiKey": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "WorkflowBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Hardened REST API for OpenMed text analysis and PII workflows.",
    "title": "OpenMed REST API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workflows/cancel": {
      "post": {
        "operationId": "workflow_cancel_v1_workflows_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernedWorkflowMutation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowView"
                }
              }
            },
            "description": "Bounded custody snapshot; no effect execution."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          }
        },
        "security": [
          {
            "WorkflowApiKey": []
          },
          {
            "WorkflowBearer": []
          }
        ],
        "summary": "Workflow Cancel",
        "tags": [
          "Governed workflows"
        ],
        "x-max-request-bytes": 65536,
        "x-required-scope": "workflow:cancel"
      }
    },
    "/v1/workflows/preflight": {
      "post": {
        "operationId": "workflow_preflight_v1_workflows_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernedWorkflowReference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowView"
                }
              }
            },
            "description": "Bounded custody snapshot; no effect execution."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          }
        },
        "security": [
          {
            "WorkflowApiKey": []
          },
          {
            "WorkflowBearer": []
          }
        ],
        "summary": "Workflow Preflight",
        "tags": [
          "Governed workflows"
        ],
        "x-max-request-bytes": 65536,
        "x-required-scope": "workflow:read"
      }
    },
    "/v1/workflows/preview": {
      "post": {
        "operationId": "workflow_preview_v1_workflows_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernedWorkflowReference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowView"
                }
              }
            },
            "description": "Bounded custody snapshot; no effect execution."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          }
        },
        "security": [
          {
            "WorkflowApiKey": []
          },
          {
            "WorkflowBearer": []
          }
        ],
        "summary": "Workflow Preview",
        "tags": [
          "Governed workflows"
        ],
        "x-max-request-bytes": 65536,
        "x-required-scope": "workflow:read"
      }
    },
    "/v1/workflows/review-receipts": {
      "post": {
        "operationId": "workflow_review_receipts_v1_workflows_review_receipts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernedWorkflowReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowView"
                }
              }
            },
            "description": "Bounded custody snapshot; no effect execution."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          }
        },
        "security": [
          {
            "WorkflowApiKey": []
          },
          {
            "WorkflowBearer": []
          }
        ],
        "summary": "Workflow Review Receipts",
        "tags": [
          "Governed workflows"
        ],
        "x-max-request-bytes": 65536,
        "x-required-scope": "workflow:review"
      }
    },
    "/v1/workflows/status": {
      "post": {
        "operationId": "workflow_status_v1_workflows_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GovernedWorkflowReference"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowView"
                }
              }
            },
            "description": "Bounded custody snapshot; no effect execution."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GovernedWorkflowError"
                }
              }
            },
            "description": "Fixed workflow or existing authentication diagnostic."
          }
        },
        "security": [
          {
            "WorkflowApiKey": []
          },
          {
            "WorkflowBearer": []
          }
        ],
        "summary": "Workflow Status",
        "tags": [
          "Governed workflows"
        ],
        "x-max-request-bytes": 65536,
        "x-required-scope": "workflow:read"
      }
    }
  }
}